requirements.txt
elasticsearch==7.10.1
PyJWT==2.1.0
python-dotenv==0.17.1
